BAQE-536: Fix possible dereference of missing Optional

+++ kie-wb-common-dmn/kie-wb-common-dmn-client/src/main/java/org/kie/workbench/common/dmn/client/property/dmn/DefaultValueUtilities.java
@@ -70,7 +70,7 @@ public class DefaultValueUtilities {
         }
     }
 
-    public static long getMaxUnusedIndex(final Collection<String> values,
+    public static int getMaxUnusedIndex(final Collection<String> values,
                                          final String prefix) {
         int maxIndex = 0;
         for (final String value : values) {
